# refactor: Improve performance for methods "formatByte" and "fromString"