Statistics for 1 000 000 bytes :
 - BytesUtils.formatByte : 17 ms VS 1000 ms => 98,3 %
 - BytesUtils.fromString : 35 ms VS  100 ms => 65,0 %